+++ beaker/services/experiment.py
@@ -51,11 +51,17 @@ class ExperimentClient(ServiceClient):
             return _get(experiment)
         except ExperimentNotFound:
             if "/" not in experiment:
-                # Now try with adding the account name.
+                # Try with adding the account name.
                 try:
                     return _get(f"{self.beaker.account.name}/{experiment}")
                 except ExperimentNotFound:
                     pass
+
+                # Try searching the default workspace.
+                if self.config.default_workspace is not None:
+                    matches = self.beaker.workspace.experiments(match=experiment, limit=1)
+                    if matches:
+                        return matches[0]
             raise
 
     def create(