+++ beaker/services/image.py
@@ -39,11 +39,17 @@ class ImageClient(ServiceClient):
             return _get(image)
         except ImageNotFound:
             if "/" not in image:
-                # Now try with adding the account name.
+                # Try with adding the account name.
                 try:
                     return _get(f"{self.beaker.account.name}/{image}")
                 except ImageNotFound:
                     pass
+
+                # Try searching the default workspace.
+                if self.config.default_workspace is not None:
+                    matches = self.beaker.workspace.images(match=image, limit=1)
+                    if matches:
+                        return matches[0]
             raise
 
     def create(