+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Fixed
+
+- Improved `Beaker.(dataset|experiment|image).get()` when looking up objects in the default workspace that were created by other users.
+
 ## [v1.12.0](https://github.com/allenai/beaker-py/releases/tag/v1.12.0) - 2022-11-23
 
 ### Added